signUp complete

+++ clint/src/screen/SignUpScreen.js
@@ -9,24 +9,35 @@ import { Store } from '../Store';
 import { toast } from 'react-toastify';
 import {getError} from '../utils';
 
-function SignInScreen() {
+
+
+function SignUpScreen() {
     const navigate = useNavigate();
     const { search } = useLocation();
     const redirectUrl = new URLSearchParams(search).get('redirect');
     const redirect = redirectUrl ? redirectUrl : '/';
+    const [name,setName] = useState('');
     const [email,setEmail] = useState('');
+    const [phone, setPhone] = useState('');
     const [password, setPassword] = useState('')
-    const { state, dispatch:ctxDispatch } = useContext(Store)
-    const {userInfo} = state;
+    const [confirmPassword, setConfirmPassword] = useState('');
+    const { state, dispatch: ctxDispatch } = useContext(Store)
+    const { userInfo } = state;
+    
 
-    const submitHendler = async(e) => {
+    const submitHendler = async (e) => {
         e.preventDefault();
+        if(password !== confirmPassword){
+            toast.error('password do not match');
+            return;
+        }
         try{
-            const { data } = await Axios.post('/api/users/signin', {
+            const { data } = await Axios.post('/api/users/signup', {
+                name,
                 email,
-                password
+                phone,
+                password,
             });
-            //console.log(data);
             ctxDispatch({type:'USER_SIGNIN', payload: data});
             localStorage.setItem('userInfo', JSON.stringify(data));
             navigate(redirect || '/');
@@ -35,11 +46,11 @@ function SignInScreen() {
             toast.error(getError(err));
             
         }
-    }
+    };
 
     useEffect(() => {
         if(userInfo){
-            navigate(redirect)
+            navigate(redirect);
         }
 
     },[navigate,redirect,userInfo]);
@@ -47,24 +58,36 @@ function SignInScreen() {
   return (
     <div>
         <Container className='small-container' >
-            <h1>Sign In</h1>
+            <h1>Sign Up</h1>
             <Form onSubmit={submitHendler} >
+                <Form.Group className="md-3" controlId="name"  >
+                    <Form.Label>Name</Form.Label>
+                    <Form.Control type='text' onChange={(e) => setName(e.target.value)} required />
+                </Form.Group>
                 <Form.Group className="md-3" controlId="email"  >
                     <Form.Label>Email</Form.Label>
                     <Form.Control type="email" onChange={(e) => setEmail(e.target.value)} required />
+                </Form.Group>
+                <Form.Group className="md-3" controlId="phone"  >
+                    <Form.Label>Phone Number</Form.Label>
+                    <Form.Control type='text' onChange={(e) => setPhone(e.target.value)} required />
                 </Form.Group>
                 <Form.Group className="md-3" controlId="password"  >
                     <Form.Label>Password</Form.Label>
                     <Form.Control type="password" onChange={(e) => setPassword(e.target.value)} required />
                 </Form.Group>
+                <Form.Group className="md-3" controlId="confirmpassword"  >
+                    <Form.Label> Confirm Password</Form.Label>
+                    <Form.Control type="password" onChange={(e) => setConfirmPassword(e.target.value)} required />
+                </Form.Group>
                 <div md={3}>
-                    <Button type='submit'>
-                        SignIn
+                    <Button type='submit' >
+                        SignUp
                     </Button>
                 </div>
                 <div md={3}>
-                    New Customer {" "}
-                    <Link to={`/signup?redirect=${redirect}`}>Create Your Account</Link>
+                    Already Have an account ? {" "}
+                    <Link to={`/signin?redirect=${redirect}`} >Sign In</Link>
                 </div>
             </Form>
         </Container>
@@ -72,4 +95,4 @@ function SignInScreen() {
   )
 }
 
-export default SignInScreen
+export default SignUpScreen
